[#67] RowのViewを新規ファイルに切り出した
