+++ contrail/Sources/View/AllRides/AllRidesScreen.swift
@@ -22,37 +22,6 @@ struct AllRidesScreen: View {
     }
 }
 
-struct AllRideRowView: View {
-    let rideDetail: RideDetail
-
-    var body: some View {
-        VStack {
-            HStack(alignment: .lastTextBaseline) {
-                HStack(alignment: .lastTextBaseline) {
-                    Text(rideDetail.distanceText)
-                        .font(.title)
-                    Text("km")
-                        .font(.subheadline)
-                }
-                .foregroundColor(.contrailBrand)
-                .padding(.vertical, 5)
-
-                Spacer()
-
-                Text(rideDetail.dateText)
-                    .font(.subheadline)
-                    .foregroundColor(.contrailLightGray2)
-                    .padding(.vertical, 5)
-            }
-            Divider()
-                .padding(.bottom, 4)
-
-            Text(rideDetail.durationText)
-            Spacer()
-        }
-    }
-}
-
 struct AllRidesScreen_Previews: PreviewProvider {
     static var previews: some View {
         AllRidesScreen()
